# docs/swagger.yaml
definitions:
...
  dbmodels.Districts:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dbmodels.Villages:
    properties:
      created_at:
        description: SubAreas            []SubArea `gorm:"many2many:sub_areas_villages;"`
        type: string
      district_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      village_code:
        description: "CityId\t\t\tint8\t`gorm:\"column:city_id\" json:\"city_id\"`\nDistrictCode\tstring\t`gorm:\"column:district_code\" json:\"district_code\"`"
        type: string
    type: object
...
  models.AttendancePostReq:
    properties:
      attendance_category:
        type: string
      attendance_category_type:
        type: string
      latitude:
        type: string
      location:
        type: string
      longitude:
        type: string
      notes:
        type: string
      photo_accuration:
        type: string
      photo_selfie:
        type: string
      sales_id:
        type: integer
      sales_name:
        type: string
      sales_phone:
        type: string
      time:
        type: string
      type_attendance:
        type: string
      type_timezone:
        type: string
    type: object
  models.BusinessTypeListBySRReq:
    properties:
      sales_type_id:
        items:
          type: integer
        type: array
    type: object
  models.BusinessTypeListRes:
    properties:
      tipe_bisnis:
        items:
          $ref: '#/definitions/models.DataModule'
        type: array
    type: object
...
  models.CallPlanActionDeleteReq:
    properties:
      call_plan_action_id:
        type: integer
    type: object
...
  models.ChangePhotoPinKTPRes:
    properties:
      description_code:
        type: string
      response_code:
        type: string
    type: object
...
  models.CheckProfilSalesRes:
    properties:
      area_aquisitions:
        $ref: '#/definitions/models.ProfileAreaAquisitions'
        type: object
      description:
        type: string
      description_code:
        type: string
      email:
        type: string
      phone:
        type: string
      photo:
        type: string
      position:
        $ref: '#/definitions/models.ProfilePosition'
        type: object
      response_code:
        type: string
      sales_id:
        type: string
      sales_name:
        type: string
      session_token:
        type: string
      sfa_id:
        type: string
      status:
        type: string
      sum_merchant:
        type: integer
      sum_verified:
        type: integer
    type: object
...
  models.DataModule:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  models.District:
    properties:
      districts:
        items:
          $ref: '#/definitions/dbmodels.Districts'
        type: array
    type: object
...
  models.LocationsRes:
    properties:
      district:
        type: string
      village:
        type: string
    type: object
...
  models.MetaData:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Ok
        type: string
      status:
        example: true
        type: boolean
    type: object
  models.ProfileAreaAquisitions:
    properties:
      city:
        type: string
      locations:
        items:
          $ref: '#/definitions/models.LocationsRes'
        type: array
      province:
        type: string
    type: object
  models.ProfilePosition:
    properties:
      area:
        type: string
      area_id:
        type: string
      branch:
        type: string
      branch_id:
        type: string
      branch_office:
        type: string
      id:
        type: string
      region:
        type: string
      region_id:
        type: string
      role:
        type: string
      role_id:
        type: string
      sub_area:
        type: string
      sub_area_id:
        type: string
    type: object
...
  models.ReportBySalesReq:
    properties:
      date_from:
        type: string
      date_to:
        type: string
      phone:
        type: string
    type: object
  models.Response:
    properties:
      data:
        type: object
      meta:
        $ref: '#/definitions/models.MetaData'
        type: object
    type: object
  models.ReverseQRRes:
    properties:
      nmid:
        type: string
    type: object
...
  models.Village:
    properties:
      villages:
        items:
          $ref: '#/definitions/dbmodels.Villages'
        type: array
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /v2.2/business_type/list:
    get:
      consumes:
      - application/json
      description: Business Type List
      operationId: Business Type List v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Business Type List Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BusinessTypeListRes'
              type: object
      summary: Business Type List
      tags:
      - OTTO SFA
...
  /v2.2/district/:city_id:
    get:
      consumes:
      - application/json
      description: District
      operationId: District
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: District Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.District'
              type: object
      summary: District
      tags:
      - OTTO SFA
  /v2.2/merchants/acquitision/dropdown/list:
    post:
      consumes:
      - application/json
      description: Acquitisions Dropdown List
      operationId: Acquitisions Dropdown List
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Acquitisions Dropdown List Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ReverseQRRes'
              type: object
      summary: Acquitisions Dropdown List
      tags:
      - OTTO SFA
...
  /v2.2/merchants/report_by_sales:
    post:
      consumes:
      - application/json
      description: ReportBySales
      operationId: ReportBySales V2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.ReportBySalesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Response'
              type: object
      summary: ReportBySales
      tags:
      - OTTO SFA
...
  /v2.2/ottosfa/merchants/change_status:
    post:
      consumes:
      - application/json
      description: Change Status Merchant
      operationId: Change Status Merchant v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Change Status Merchant Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.District'
              type: object
      summary: Change Status Merchant
      tags:
      - OTTO SFA
...
  /v2.2/sales/profile:
    get:
      consumes:
      - application/json
      description: Profile Sales
      operationId: Profile Sales
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile Sales Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CheckProfilSalesRes'
              type: object
      summary: Profile Sales
      tags:
      - OTTO SFA
  /v2.2/sfa/attendance/post:
    post:
      consumes:
      - application/json
      description: Attendance Post
      operationId: Attendance Post v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.AttendancePostReq'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Post Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ChangePhotoPinKTPRes'
              type: object
      summary: Attendance Post
      tags:
      - OTTO SFA
...
  /v2.2/sfa/call_plan/action/delete:
    post:
      consumes:
      - application/json
      description: CallPlan Action Delete
      operationId: CallPlan Action Delete
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CallPlanActionDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: CallPlan Action Delete Response EXAMPLE
          schema:
            $ref: '#/definitions/models.Response'
      summary: CallPlan Action Delete
      tags:
      - OTTO SFA
...
  /v2.2/village/:district_id:
    get:
      consumes:
      - application/json
      description: Village
      operationId: Village
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Village Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Village'
              type: object
      summary: Village
      tags:
      - OTTO SFA
...
  /v2.3/merchants/business_type/list:
    get:
      consumes:
      - application/json
      description: Business Type List
      operationId: Business Type List v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Business Type List Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BusinessTypeListRes'
              type: object
      summary: Business Type List
      tags:
      - OTTO SFA
    post:
      consumes:
      - application/json
      description: Business Type List  By SR
      operationId: Business Type List  By SR
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.BusinessTypeListBySRReq'
      produces:
      - application/json
      responses:
        "200":
          description: Business Type List By SR Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BusinessTypeListRes'
              type: object
      summary: Business Type List  By SR
      tags:
      - OTTO SFA
...
  /v2.3/sfa/attendance/post:
    post:
      consumes:
      - application/json
      description: Attendance Post
      operationId: Attendance Post v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.AttendancePostReq'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Post Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ChangePhotoPinKTPRes'
              type: object
      summary: Attendance Post
      tags:
      - OTTO SFA
...
swagger: "2.0"